Fix issue with root command inheritance
In attempting to avoid using global variables in the way cobra
encourages you to, we were failing to pick up root command flags when
supplied due to them being populated after the point at which they were
required. This prevented me from using the CLI tool against a local
console deployment.

To address the issue, child commands are initialized with pointers to
the flag values instead of copies of them.