Fix deadlock in stdout redirection test (#19)
This uncomments a test that was previously commented out locally and
mistakenly committed. The reason the test was commented out was a
deadlock in one of the test utilities, which I believe was caused by
stdout getting redirected too late, after an earlier reference was taken
by the writer. This commit does a bit of rearranging to avoid the
deadlock.