Add "devices" command with "list" subcommand (#33)
Adds a new top-level "devices" command, under which we can place device
management and interogation subcommands. The first subcommand
implemented in this patch is "list".

The "foxglove devices list" call supports an optional parameter,
"format", which can be used to indicate the desired output format for
the list of devices. Supported formats are "table", "json", and "csv".
The default format is "table".