Enable streaming by import ID (#79)
* Enable exports by import ID

Adds support for an --import-id flag to the export command, to allow
users to download data by import ID.

* Stop returning key in imports response

Changes the "Import ID" column in the imports response listing to point
at import short hashes rather than the previous content ID key. This
short hash is what will be used for streaming by import ID.

* Fix lint error

* Add validation for export requests

Validate that the end parameter is after the start parameter when
supplied, and clarify the validation failure if import ID and device ID
are both unspecified.