Import: validate input file appears legal (#87)
This commit performs a sanity check on input bags or mcap files to
ensure that the magic bytes (leading in case of bag, leading and
trailing in case of mcap) match expected. The main benefit of this is
detection of empty files or truncated mcap files prior to ingestion into
the data platform.

It does not parse the entire input, decompress any chunks, or validate
anything beyond magic - errors resulting from bad content will be
detected once the input gets to the backend. If full validation is
required, running the file through `rosbag` or `mcap doctor` can be
performed as a separate step, at the cost of some time (for large inputs
in particular).