Support resumable exports across broken downloads (#95)
* Add ROS reader and writer

Adds a ROS bag lexer and bag writer, similar in function and usage to
the ones in the MCAP repo. Produces fully indexed and ROS-compatible
bags.

* Add resumable download support for the export command

Introduces a new --output-file arguement to the export command, which
previously output to stdout. When --output-file is supplied, resumable
download behavior is triggered (this is not supportable when downloading
to stdout).

When using the resumable download behavior, the export is initially
written to a temporary directory. When the export finishes, the tool
does some checks on the download to ensure it is complete. If the
download is not complete, additional downloads are requested, and the
exports are merged at the end.

The logic used depends on whether the file is an MCAP or a bag file.
MCAP files can be easily checked for completeness based on the presence
of closing MCAP magic, whereas with an unindexed ROS bag it is never
clear whether the file is complete or has been truncated on a record
boundary.

Accordingly, for MCAP files the process is
1. Download to temporary location
2. Check validity of the index
3. If index is not valid, reindex the file and issue a new request
   starting on the final timestamp of the file.
4. Repeat until a valid index is found, then stop

For Bag files it is,
1. Download to a temporary location
2. Reindex the bag
3. Make a new request starting at the largest timestamp
4. If two successive requests are made starting with the same timestamp,
   assume we have found the end of the data, and stop.

In both cases, we then merge the partial outputs, excluding the final
timestamp in the files except in the case of the last output.

This comes with some caveats for bag files, even when there is no broken
download:
* We always pay the cost of reindexing the download. In other words,
  scanning the full file.
* We always make a couple more requests than we need, to convince
  ourselves we are at the end.

This means performance will be worse on bags than on mcap, and also it
will be worse on bags than it is with the stdout output option. On the
plus side, bag exports after this change are fully indexed.

* Add BufWriteSeeker

* add bag reader

* writer fixes

* Address underflow

* Remove Close

* Add docstrings