Add tab autocompletion for device names (#109)
* Add tab autocompletion for device names

* Make autocompletion work for alternative base URLs

This changes the way the --base-url parameter works. It can now only be
suppplied to "foxglove auth login" and "foxglove auth
configure-api-key". The base URL will be written to the config file, and
read from there by other commands. This is backward compatible so as not
to break existing calls. If there is no base_url in that file, the
default production URL will be assumed.

As side effects,
* it will no longer be necessary to specify --base-url to every command
  issued.
* autocompletion features will work for the configured base URL