export: fix bag reindexing (#141)
### Changelog

- Fixed: previously ROS 1 bag files downloaded using the foxglove CLI
would fail with an "unindexed bag" error. These files should now
download correctly.

### Docs


### Description

Since #124, the code to reindex exported bag files has been broken - it
first tries to use a seeking reader to read the unindexed file (which
will not work), then it fails to close the reindexing writer in the
happy path. This PR fixes that and adds a test.