Projects support (#145)
### Changelog

Introduces the concept of Projects in the CLI.

### Docs

https://github.com/foxglove/docs/pull/832

### Description

Introduces the concept of
[projects](https://docs.foxglove.dev/docs/projects).

First, the new `foxglove projects list` feature will return the list of
projects available to the user:

```
|          ID          |      Name      |  Member Count  |  Last Recording Uploaded At  |
|----------------------|----------------|----------------|------------------------------|
| prj_0dixnOnBJwRLfXqI | first project  | 1              | 2025-06-30T16:11:04Z         |
| prj_0dj0oByUzy58EUxp | second project | 2              |                              |
| prj_0dj0oE74rIswv7MS | third project  | 3              | 2025-06-26T16:01:53Z         |
```

Using the returned project IDs we can also upload data to a specific
project, and filter the existing recordings by the project. This PR
prints the `Project ID` column to all endpoints that return it, and adds
support for the `--project-id` parameter as required:

- `foxglove devices list --project-id <projectId>`
- `foxglove devices add --name <name> --project-id <projectId>`
- `foxglove recordings list --project-id <projectId>`
- `foxglove pending-imports list --project-id <projectId>`
- `foxglove data coverage list --project-id <projectId>`
- `foxglove data imports add --project-id <projectId>`

The following features neither accept the `projectId` parameter, nor
return the field in the response:

- `foxglove attachments list`
- `foxglove events list`
- `foxglove data import` (This is a *deprecated* endpoint and we should
consider removing it from the CLI)

Note that the `projectId` is required if-and-only-if where the request
is ambiguous without it. For example, uploading data to an existing
device doesn't require a project ID, because the backend can determine
it using the device ID. Where `projectId` is required but isn't
provided, the CLI will forward the API error ("Single project was not
found...")

Other features considered:
- Users can create and delete projects through the website, where we
display a list of notes/documentation which they really should read
(especially before creating their "first" project). We're not
implementing these features here (yet)
- Similarly, managing members who can access a project is outside the
scope. (The CLI is scoped for data management; for example, admins need
to use the website for inviting organization members today.)
- Saving the default project ID coming up separately in
[FG-11994](https://linear.app/foxglove/issue/FG-11994/support-projects-in-cli)